Refactor StepCard from WalletBrowser as separate component
